ft: S3C-1065: handle null parts on sparse objects
(null parts are parts with null keys which are generated by the NFS server when
growing a file with truncate)

Because of the data/wrapper batchDelete function becoming 'more' asynchronous
by calling data.delete from nextTick (that prevents a call stack overflow on
large array), some of the process.nextTick calls in the unit tests have been
replaced by setImmediate. (setImmediate is scheduled AFTER nextTick)

As well, This change clarifies the predicates about the md-model-version and the
backends string requirements.